# docs/swagger.yaml
basePath: /api/v1
definitions:
  types.ApiAlbum:
    properties:
      artistId:
        type: string
      coverArt:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  types.ApiArtist:
    properties:
      id:
        type: string
      name:
        type: string
      picture:
        type: string
    type: object
  types.ApiError:
    properties:
      data: {}
      message:
        type: string
      status:
        example: 400
        type: integer
    type: object
  types.ApiGetAlbumByIdData:
    properties:
      artistId:
        type: string
      coverArt:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  types.ApiGetAlbumTracksByIdData:
    properties:
      tracks:
        items:
          $ref: '#/definitions/types.ApiTrack'
        type: array
    type: object
  types.ApiGetAlbumsData:
    properties:
      albums:
        items:
          $ref: '#/definitions/types.ApiAlbum'
        type: array
    type: object
  types.ApiGetArtistAlbumsByIdData:
    properties:
      albums:
        items:
          $ref: '#/definitions/types.ApiAlbum'
        type: array
    type: object
  types.ApiGetArtistByIdData:
    properties:
      id:
        type: string
      name:
        type: string
      picture:
        type: string
    type: object
  types.ApiGetArtistsData:
    properties:
      artists:
        items:
          $ref: '#/definitions/types.ApiArtist'
        type: array
    type: object
  types.ApiGetTrackById:
    properties:
      albumId:
        type: string
      albumName:
        type: string
      artistId:
        type: string
      artistName:
        type: string
      bestQualityFile:
        type: string
      coverArt:
        type: string
      id:
        type: string
      mobileQualityFile:
        type: string
      name:
        type: string
      number:
        type: integer
    type: object
  types.ApiGetTracksData:
    properties:
      tracks:
        items:
          $ref: '#/definitions/types.ApiTrack'
        type: array
    type: object
  types.ApiPostAlbumData:
    properties:
      artistId:
        type: string
      coverArt:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  types.ApiPostArtistData:
    properties:
      id:
        type: string
      name:
        type: string
      picture:
        type: string
    type: object
  types.ApiPostTrackData:
    properties:
      albumId:
        type: string
      albumName:
        type: string
      artistId:
        type: string
      artistName:
        type: string
      bestQualityFile:
        type: string
      coverArt:
        type: string
      id:
        type: string
      mobileQualityFile:
        type: string
      name:
        type: string
      number:
        type: integer
    type: object
  types.ApiResponse-types_ApiGetAlbumByIdData:
    properties:
      data:
        $ref: '#/definitions/types.ApiGetAlbumByIdData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiGetAlbumTracksByIdData:
    properties:
      data:
        $ref: '#/definitions/types.ApiGetAlbumTracksByIdData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiGetAlbumsData:
    properties:
      data:
        $ref: '#/definitions/types.ApiGetAlbumsData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiGetArtistAlbumsByIdData:
    properties:
      data:
        $ref: '#/definitions/types.ApiGetArtistAlbumsByIdData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiGetArtistByIdData:
    properties:
      data:
        $ref: '#/definitions/types.ApiGetArtistByIdData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiGetArtistsData:
    properties:
      data:
        $ref: '#/definitions/types.ApiGetArtistsData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiGetTrackById:
    properties:
      data:
        $ref: '#/definitions/types.ApiGetTrackById'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiGetTracksData:
    properties:
      data:
        $ref: '#/definitions/types.ApiGetTracksData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiPostAlbumData:
    properties:
      data:
        $ref: '#/definitions/types.ApiPostAlbumData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiPostArtistData:
    properties:
      data:
        $ref: '#/definitions/types.ApiPostArtistData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiResponse-types_ApiPostTrackData:
    properties:
      data:
        $ref: '#/definitions/types.ApiPostTrackData'
      status:
        example: 200
        type: integer
    type: object
  types.ApiTrack:
    properties:
      albumId:
        type: string
      albumName:
        type: string
      artistId:
        type: string
      artistName:
        type: string
      bestQualityFile:
        type: string
      coverArt:
        type: string
      id:
        type: string
      mobileQualityFile:
        type: string
      name:
        type: string
      number:
        type: integer
    type: object
info:
  contact: {}
  title: Testing Testing
paths:
  /albums:
    get:
      description: Get all albums
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiGetAlbumsData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get all albums
      tags:
      - albums
    post:
      consumes:
      - multipart/form-data
      description: Create new album
      parameters:
      - description: Album name
        in: formData
        name: name
        required: true
        type: string
      - description: Artist Id
        in: formData
        name: artistId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiPostAlbumData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Create new album
      tags:
      - albums
  /albums/{id}:
    get:
      description: Get album by id
      parameters:
      - description: Album Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiGetAlbumByIdData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get album by id
      tags:
      - albums
  /albums/{id}/tracks:
    get:
      description: Get all tracks from album
      parameters:
      - description: Artist Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiGetAlbumTracksByIdData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get all tracks from album
      tags:
      - albums
  /artists:
    get:
      description: Get all artists
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiGetArtistsData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get all artists
      tags:
      - artists
    post:
      consumes:
      - multipart/form-data
      description: Create new artist
      parameters:
      - description: Artist name
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiPostArtistData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Create new artist
      tags:
      - artists
  /artists/{id}:
    get:
      description: Get artist by id
      parameters:
      - description: Artist Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiGetArtistByIdData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get artist by id
      tags:
      - artists
  /artists/{id}/albums:
    get:
      description: Get all albums by artist
      parameters:
      - description: Artist Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiGetArtistAlbumsByIdData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get all albums by artist
      tags:
      - artists
  /tracks:
    get:
      description: Get all tracks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiGetTracksData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get all tracks
      tags:
      - tracks
    post:
      consumes:
      - multipart/form-data
      description: Create new track
      parameters:
      - description: Track name
        in: formData
        name: name
        required: true
        type: string
      - description: Track number
        in: formData
        name: number
        required: true
        type: string
      - description: Album Id
        in: formData
        name: albumId
        required: true
        type: string
      - description: Artist Id
        in: formData
        name: artistId
        required: true
        type: string
      - description: Best Quality File
        in: formData
        name: bestQualityFile
        required: true
        type: file
      - description: Mobile Quality File
        in: formData
        name: mobileQualityFile
        required: true
        type: file
      - description: Cover Art
        in: formData
        name: coverArt
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiPostTrackData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Create new track
      tags:
      - tracks
  /tracks/{id}:
    get:
      description: Get track by id
      parameters:
      - description: Track Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.ApiResponse-types_ApiGetTrackById'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ApiError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/types.ApiError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/types.ApiError'
      summary: Get track by id
      tags:
      - tracks
swagger: "2.0"